Expand serialization_specs recursively Also extract tests into different files
